Path: add incremental utility functions (from delta fixes, updates)
Added:
- IsLocalWindows() (consolidate from multiple places)
- ToBareLocal(), ToFileURI() - for converting between / (bare) and file:///
- Add GetSchemeKind && scheme kind enums, giving UNC and family `Verbatim` name (also used in rust fs api)
- add ToDot() to clarify '.' normalizations and provide a checker
- "relative" funcs, as seen in python pathlib (our main source of inspiration), and c++17
    - add IsRelativeTo() for parentage checks, a la Path("/a/b/c.txt").IsRelativeTo("/a/b") -> true
    - add RelativeTo() to produce relative path when appropriate
- add Equals, EqualsCI, and EqualsCanonical
- add simple ToCanonical -- requires explicit base absolute path e.g., $CWD; can
  consider a variant which auto fills with $CWD later

Also:
- moved cpp unit tests from test_file_system to test_path.cpp
- leverage IsRelativeTo internally in Join checks
