+++ src/include/duckdb/common/path.hpp
@@ -17,15 +17,15 @@ namespace duckdb {
 // Parsed representation of a path string, covering posix, windows, URI, and UNC forms.
 //
 // FromString(raw) parses and normalizes the input; ToString() reconstructs it as
-// scheme + authority + anchor + join(segments, sep) + tail
+// scheme + authority + anchor + join(segments, sep) + trailing_sep
 //
 // Supported input forms and their parsed fields:
 //
 //   Input                         scheme        authority      anchor   segments      IsAbsolute
 //   ----------------------------  ------------  -----------    -------  ------------  -----------
 //   "a/b"                         ""            ""             ""       "a/b"         false
-//   ""  (empty)                   ""            ""             ""       "."           false
-//   "/"                           ""            ""             "/"      ""            true
+//   ""  (empty)                   ""            ""             ""       ""            false
+//   "/"                           ""            ""             "/"      ""            true   * see below
 //   "/a/b"                        ""            ""             "/"      "a/b"         true
 //   "/a/b/"                       ""            ""             "/"      "a/b"         true   (HasTrailingSeparator)
 //   "file:/a/b"                   "file:"       ""             "/"      "a/b"         true
@@ -38,6 +38,8 @@ namespace duckdb {
 //   "\\?\UNC\server\share" (win)  "\\?\UNC\"    "server\share" "\"      ""            true
 //   "\\?\C:\foo" (win)            "\\?"         ""             "C:\"    "foo"         true
 //
+// *On empty Paths: Path("") == Path("."), Path("").IsDot() == true, and Path("").ToString == "."
+//
 // HasTrailingSeparator is true when raw input ends with a separator. ToString() re-emits the
 // separator after the last segment (skipped when segments is empty — anchor already ends with it).
 // When Join()ing, LHS inherits trailing_separator from RHS. This is useful (and semantically
@@ -48,8 +50,20 @@ namespace duckdb {
 //
 class Path {
 public:
-	// Primary Constructor
-	static Path FromString(const string &raw);
+	enum class SchemeKind {
+		None,     // no scheme — plain /foo, C:\foo, relative paths
+		File,     // file: prefix — file:/foo, file:///foo, file://localhost/foo
+		UNC,      // network UNC — \\server\share\... (plain or \\?\UNC\ pre-canonical)
+		URI,      // remote URI — s3://, az://, http://, etc.
+		Verbatim, // \\?\ extended-length prefix (Windows) — collapses to None or UNC on ToCanonical()
+	};
+
+	// Primary Constructor. If canonicalize=true, applies ToCanonical() after parsing (requires absolute).
+	static Path FromString(const string &raw, bool canonicalize = false);
+
+	Path() = default;
+	explicit Path(const string &s, bool canonicalize = false) : Path(FromString(s, canonicalize)) {
+	}
 
 	string ToString() const;
 
@@ -82,7 +96,7 @@ public:
 		return !anchor.empty();
 	}
 
-	bool HasDrive() const;     // always false in non-windows
+	bool HasDrive() const;     // false unless anchor contains a drive letter; currently only set by the Windows parser
 	char GetDriveChar() const; // returns \0 if no drive
 
 	bool HasPathSegments() const {
@@ -93,14 +107,53 @@ public:
 		return has_trailing_separator;
 	}
 
+	// true for all network paths, /*,  c:/* (not c:foo), file:/*, \\?\C:\*
 	bool IsAbsolute() const {
 		return is_absolute;
 	}
 
-	// true for all relative paths, /*,  c:/* (not c:foo), file:/*, \\?\C:\*
+	bool IsRelative() const {
+		return !is_absolute;
+	}
+
+	// True iff this is the "current directory" relative path: no scheme, authority, anchor, or segments.
+	// Both "" and "." normalize to this; Path("a/..") does too after normalization.
+	// Path(".").Equals(Path()) == Path("").Equals(Path(".")) == true.
+	bool IsDot() const {
+		return !is_absolute && scheme.empty() && authority.empty() && anchor.empty() && segments.empty();
+	}
+
+	SchemeKind GetSchemeKind() const;
+
+	bool IsNoneScheme() const {
+		return GetSchemeKind() == SchemeKind::None;
+	}
+	bool IsFileScheme() const {
+		return GetSchemeKind() == SchemeKind::File;
+	}
+	bool IsUNCScheme() const {
+		return GetSchemeKind() == SchemeKind::UNC;
+	}
+	bool IsURIScheme() const {
+		return GetSchemeKind() == SchemeKind::URI;
+	}
+	bool IsVerbatimScheme() const {
+		return GetSchemeKind() == SchemeKind::Verbatim;
+	}
+
+	// Local: None (all bare paths), File scheme, or Verbatim with a drive letter (\\?\C:\).
+	// Verbatim UNC (\\?\UNC\), plain UNC, and URI schemes are all remote.
 	bool IsLocal() const {
-		// note: HasDrive() covers UNC locals too!
-		return (!IsAbsolute() || !HasScheme() || HasDrive() || StringUtil::StartsWith(scheme, "file:"));
+		switch (GetSchemeKind()) {
+		case SchemeKind::None:
+			return true;
+		case SchemeKind::File:
+			return true;
+		case SchemeKind::Verbatim:
+			return HasDrive(); // \\?\C:\ is local; \\?\UNC\ is not
+		default:
+			return false;
+		}
 	}
 
 	bool IsRemote() const {
@@ -111,7 +164,10 @@ public:
 		return segments;
 	}
 
-	// Join (in several forms)
+	// Join rhs onto this path:
+	//   - relative RHS: appends to LHS (resolving any .. against LHS segments)
+	//   - absolute RHS: returned as-is IFF RHS.IsRelativeTo(LHS) (i.e. RHS extends LHS within same root)
+	//   - all others: incompatible, throws InvalidInputException
 	Path Join(const Path &rhs) const;
 	Path Join(const string &rhs) const;
 
@@ -128,7 +184,32 @@ public:
 		return result;
 	}
 
-	Path Parent(int n = 1) const;
+	Path Parent(int n = 1) const; // apply ".." n times; n=0 returns self
+
+	// True if `base` is a prefix of (or equal to) this path — same scheme/authority/anchor required.
+	bool IsRelativeTo(const Path &base) const;
+	// Returns the suffix of this path after stripping `base`; throws if !IsRelativeTo(base).
+	// e.g. Path("/a/b/c").RelativeTo("/a") == "b/c";  Path("/a").RelativeTo("/a") == "."
+	Path RelativeTo(const Path &base) const;
+
+	// strip scheme/authority → bare notation (no scheme); throws if non-local. Relative passes through.
+	Path ToBareLocal() const;
+	// normalize to file:///... form; throws if non-local or relative
+	Path ToFileURI() const;
+	// Canonical form (throws if relative): collapses file: variants and \\?\ extended-length prefixes
+	// to their simplest absolute form; remote paths are unchanged. Canonical implies absolute.
+	Path ToCanonical() const;
+
+	// operator== is intentionally omitted: normalization does not imply canonicalization, so naive
+	// field equality yields surprises. "file:/foo", "file:///foo", and "/foo" are structurally
+	// distinct after parsing but refer to the same path. Use one of:
+	//   Equals()         — field-by-field, case-sensitive (posix / URI)
+	//   EqualsCI()       — case-insensitive throughout (Windows local + UNC)
+	//   EqualsCanonical() — canonicalize both sides first, then Equals()
+	// bool operator==(const Path &) const; // intentionally omitted — see above
+	bool Equals(const Path &other) const;
+	bool EqualsCI(const Path &other) const;
+	bool EqualsCanonical(const Path &other) const;
 
 	// public convenience - string-to-string normalize
 	static string Normalize(const string &input) {
@@ -146,6 +227,13 @@ private:
 	char separator = '/';
 	bool has_trailing_separator = false;
 	bool is_absolute = false;
+
+	// True when path should use case-insensitive comparison: always IsLocal() on Windows;
+	// off-platform, inferred from Windows-shaped paths (drive/UNC/Verbatim) for cross-platform correctness.
+	bool IsLocalWindows() const;
+	// std::equal with case sensitivity matching the path's filesystem
+	bool SegmentsEqual(vector<string>::const_iterator a, vector<string>::const_iterator a_end,
+	                   vector<string>::const_iterator b) const;
 
 	void NormalizeSegments(const string &raw, size_t path_offset);
